requirements.txt
# Core
anthropic>=0.40.0
supabase>=2.4.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

# Email verification
# zerobounce-sdk  # Uses REST directly via httpx — no SDK needed

# Dashboard
streamlit>=1.35.0
pandas>=2.2.0

# LinkedIn scraping (Tier 2) — optional, install separately
# camoufox[geoip]
# playwright

# Dev / testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
